break provider search result into a component
